+++ inst/examples/movingmarker_app.R
@@ -3,15 +3,11 @@ library(sf)
 library(leaflet)
 library(leaflet.extras2)
 
-# df <- sf::st_as_sf(atlStorms2005)
-# df <- df[1,]
-#
-# dfp <- st_cast(df, "POINT")
-# dfp$label = sample(LETTERS[1:20], nrow(dfp), TRUE)
-# dfp$popup = sample(LETTERS[1:20], nrow(dfp), TRUE)
-# dfp$duratios = sample(c(1000, 2000, 3000, 4000, 5000), nrow(dfp), TRUE)
+df <- sf::st_as_sf(atlStorms2005)[1,]
+dfp <- st_cast(df, "POINT")
+dfp$duratios = sample(c(1000, 1500, 2000, 2500, 3000), nrow(dfp), TRUE)
 
-greenLeafIcon <- makeIcon(
+shipIcon <- makeIcon(
   iconUrl = "https://image.flaticon.com/icons/png/512/905/905567.png",
   iconWidth = 40, iconHeight = 50,
   iconAnchorX = 0, iconAnchorY = 0
@@ -26,8 +22,19 @@ ui <- fluidPage(
   actionButton("addLatLng", "addLatLng"),
   actionButton("moveTo", "moveTo"),
   actionButton("addStation", "addStation"),
-  h5("Click Events"),
-  verbatimTextOutput("click")
+  actionButton("clear", "Clear Group"),
+  actionButton("clearmark", "Clear Marker"),
+  actionButton("remove", "Remove Marker"),
+  splitLayout(
+    div(
+      h5("Click Events"),
+      verbatimTextOutput("click")
+    ),
+    div(
+      h5("Mouseover Events"),
+      verbatimTextOutput("mouseover")
+    )
+  )
 )
 
 server <- function(input, output, session) {
@@ -35,15 +42,13 @@ server <- function(input, output, session) {
     leaflet()  %>%
       addTiles() %>%
       addPolylines(data = df, group = "mapkey") %>%
-      # addMovingMarker(data = df, group = "mapkey", icon = icon("cogs"))
       addMovingMarker(data = dfp,
                       duration = ~duratios,
-                      # duration = 1000,
-                      icon = greenLeafIcon,
+                      icon = shipIcon,
                       movingOptions = movingMarkerOptions(autostart = TRUE, loop = TRUE),
-                      options = list(title="title some"),
-                      label=~label, popup=dfp$popup[1],
-                      group = "movingmarker"
+                      label="I am a pirate!",
+                      popup="Arrr",
+                      group = "movingmarker", layerId = "myid"
                       )
   })
 
@@ -51,6 +56,11 @@ server <- function(input, output, session) {
     txt <- req(input$map_movingmarker_click)
     print(txt)
   })
+  output$mouseover <- renderPrint({
+    txt <- req(input$map_movingmarker_mouseover)
+    print(txt)
+  })
+
   observeEvent(input$start, {
     leafletProxy("map", session) %>%
       startMoving()
@@ -68,20 +78,33 @@ server <- function(input, output, session) {
       resumeMoving()
   })
 
-  # observeEvent(input$addLatLng, {
-  #   leafletProxy("map", session) %>%
-  #     addLatLngMoving(latlng, duration)
-  # })
-  # observeEvent(input$moveTo, {
-  #   leafletProxy("map", session) %>%
-  #     moveToMoving(latlng, duration)
-  # })
+  observeEvent(input$addLatLng, {
+    leafletProxy("map", session) %>%
+      addLatLngMoving(list(33, -67), 2000)
+  })
+  observeEvent(input$moveTo, {
+    leafletProxy("map", session) %>%
+      moveToMoving(list(lng=-65, lat=33), 2000)
+  })
   observeEvent(input$addStation, {
     pti <- sample(seq.int(nrow(dfp)), 1, TRUE)
     print(paste("Stay at Point Index:", pti, "for 4000 ms."))
     leafletProxy("map", session) %>%
       addStationMoving(pointIndex = pti,
                        duration = 4000)
+  })
+
+  observeEvent(input$clear, {
+    leafletProxy("map", session) %>%
+      clearGroup("movingmarker")
+  })
+  observeEvent(input$clearmark, {
+    leafletProxy("map", session) %>%
+      clearMarkers()
+  })
+  observeEvent(input$remove, {
+    leafletProxy("map", session) %>%
+      removeMarker("myid")
   })
 }
 shinyApp(ui, server)